==> notebooks/4 combine_measurements_and_metadata.ipynb <==
import geopandas as gpd
import pandas as pd

from fwi_predict.constants import WQ_RANGES
from fwi_predict.wq import get_in_required_range

ponds = gpd.read_file("../data/clean/pond_metadata_clean.geojson")
measurements = pd.read_csv("../data/clean/ara_measurements_clean.csv",
                           parse_dates=['sample_dt', 'prescribed_collection_date'])

measurements['sample_dt']

# Subset to columns of interest
ponds_cols = ['pond_id', 'property_area_acres', 'pond_area_acres',
                      'pond_depth_meters', 'geometry']
pond_measurement_cols = ['pond_id', 'region', 'farm_id', 'sample_dt', 'do_mg_per_L',
                         'ph', 'turbidity_cm', 'ammonia_mg_per_L', 'time_of_day']

# Construct days since last measurement variable. This would suggest including water quality correction

ponds = ponds[ponds_cols]
ponds = ponds[ponds['geometry'].notna()] # Have to have locations to get weather data.
measurements = measurements[measurements['follow_up'] == False]
measurements = measurements[pond_measurement_cols]

# Remove measurements with inconsistent times of day
measurements['tod_dt'] = measurements['sample_dt'].apply(lambda x: "morning" if x.hour < 12 else "evening")
print(f"Observations with inconsistent times: {len(measurements[measurements['tod_dt'] != measurements['time_of_day']])}")
measurements = measurements[measurements['tod_dt'] == measurements['time_of_day']]
measurements = measurements.drop(columns=['tod_dt'])

combined = gpd.GeoDataFrame(
  measurements.merge(ponds, on='pond_id', how='inner'), # Inner as we need location
  geometry='geometry',
  crs=ponds.crs
)

ponds_not_included = ponds[~ponds['pond_id'].isin(combined['pond_id'])]
print(f"No measurements for {len(ponds_not_included)} ponds")

combined.describe().T

combined['sample_idx'] = pd.Series(range(len(combined)))
combined.to_file("../data/clean/measurements_with_metadata.geojson")

# report
