==> notebooks/25 clean_stocking_records.ipynb <==
import pandas as pd
from ydata_profiling import ProfileReport

raw = pd.read_excel("../data/raw/Stocking & Harvest September 2024.xlsx")

colname_map = {
  'sr._no': 'record_idc',
  'pond': 'pond_id',
}

stocking_records = raw \
  .rename(columns=lambda s: s.lower() \
															.replace(' ', '_') \
															.replace('(', '') \
															.replace(')', '') \
															.replace(',', '') \
															.replace('?', '')) \
  .rename(columns=colname_map)

stocking_records.columns

# Throw out columns we're not going to look at
drop_cols = ['earnings_from_the_fish/shrimp_sale_in_inr',
             'how_much_did_the_farmer_pay_for_the_stocked_fish_and_shrimp_in_inr']
stocking_records = stocking_records.drop(columns=drop_cols)

stocking_records

report = ProfileReport(stocking_records)

report
