==> notebooks/30 check_most_recent_gfs.ipynb <==
from datetime import datetime

import ee
import geopandas as gpd
import pandas as pd

from fwi_predict.geo.ee import get_gfs

# Initialize Earth Engine
ee.Initialize()

# Read pond metadata
ponds = gpd.read_file("../data/clean/pond_metadata_clean.geojson")

# Get first pond coordinate to use as representative point
first_lon = ponds.geometry.iloc[0].x
first_lat = ponds.geometry.iloc[0].y
point = ee.Geometry.Point(first_lon, first_lat)

# Get GFS collection
gfs = get_gfs()

point

# Get most recent forecast date
tomorrow = ee.Date(datetime.now()).advance(1, 'day')
latest_forecast = gfs \
		.filterDate(tomorrow.advance(-15, 'day'), tomorrow) \
		.filterBounds(point) \
    .sort('system:time_start', False) \
    .first()
forecast_date = datetime.fromtimestamp(
    latest_forecast.get('system:time_start').getInfo()/1000
).astimezone()

print(f"Most recent GFS forecast available is from: {forecast_date}")
