==> notebooks/31a sentinel2_for_jan_mar_25_test.ipynb <==
from pathlib import Path

import ee
import geopandas as gpd
import pandas as pd
from geemap import gdf_to_ee

from fwi_predict.geo.ee import monitor_task, get_sentinel2_values_at_feature
from fwi_predict.gcs import download_files

ee.Initialize(project='fwi-water-quality-sensing')

predict_ds = gpd.read_file("../data/clean/jan_mar_25_w_metadata.geojson")

gcs_fp = Path('train/sentinel2/jan_mar_25_w_metadata').as_posix()

def create_sentinel2_feature_set(ds: gpd.GeoDataFrame,
                                 filename_prefix: str,
                                 bucket: str = 'fwi-predict',
																 back_days: int = 20):
	
	small_df = ds[['sample_idx', 'sample_dt', 'geometry']]
	ds_ee = gdf_to_ee(small_df, date='sample_dt', date_format="yyyy-MM-dd'T'HH:mm:ssZ")
	# ds_ee = ds_ee.map(lambda)
	fc = ds_ee.map(lambda f: get_sentinel2_values_at_feature(f, back_days=back_days))

	task = ee.batch.Export.table.toCloudStorage(
		collection=fc,
		description='mwm_sentinel2',
		bucket='fwi-predict',
		fileNamePrefix=gcs_fp,
		fileFormat='CSV'
	)
	task.start()

	task_success = monitor_task(task)

	if not task_success:
		print("Data export failed. Please consult GEE task manager for information.")
		return None
	
	download_files(bucket='fwi-predict',
                 file_glob=gcs_fp + '.csv',
                 download_dir='../data/gcs')
	
	print(f"Saving Sentinel-2 features to: ../data/gcs/{gcs_fp}.csv")


create_sentinel2_feature_set(predict_ds, filename_prefix=gcs_fp)

sentinel_features = pd.read_csv("../data/gcs/train/sentinel2/jan_mar_25_w_metadata.csv").drop(columns=['system:index', '.geo'])

sentinel_features
